+++ RELEASE_NOTES.md
@@ -7,6 +7,8 @@ Release Notes
 
 ## 1.6.4
 * Azure Firewall: Added support for Azure Firewalls
+* Service bus: Support max queue and topic sizes.
+* Service bus: Set default message TTL for subscriptions.
 
 ## 1.6.3
 * Container Service (AKS): Support basic SKU for the cluster's load balancer (default is standard).